consolidate bids.spec
